get rid from bart attention copypaste
